--- a/src/reports_updater.rs
+++ b/src/reports_updater.rs
@@ -1,11 +1,11 @@
-use crate::{networking::{TcnApi, NetworkingError}, reports_interval, DB_UNINIT, DB, byte_vec_to_16_byte_array, errors::{Error, ServicesError}, preferences::{PreferencesKey, Preferences}, byte_vec_to_24_byte_array, byte_vec_to_8_byte_array, reporting::{public_report::{PublicReport, FeverSeverity, CoughSeverity}, memo::{Memo, MemoMapper}, symptom_inputs::UserInput}};
+use crate::{networking::{TcnApi, NetworkingError}, reports_interval, DB_UNINIT, DB, byte_vec_to_16_byte_array, errors::{Error, ServicesError}, preferences::{PreferencesKey, Preferences}, byte_vec_to_24_byte_array, byte_vec_to_8_byte_array, reporting::{public_report::PublicReport, memo::{Memo, MemoMapper}}};
 use reports_interval::{ReportsInterval, UnixTime};
 use tcn::{TemporaryContactNumber, SignedReport};
 use std::{time::Instant, sync::Arc};
 use serde::Serialize;
 use uuid::Uuid;
 use chrono::Utc;
-use std::{thread, collections::HashMap, io::Cursor};
+use std::{thread, collections::HashMap};
 use thread::JoinHandle;
 use rayon::prelude::*;
 
@@ -13,98 +13,8 @@ pub trait TcnMatcher {
   fn match_reports(&self, tcns: Vec<ObservedTcn>, reports: Vec<SignedReport>) -> Result<Vec<MatchedReport>, ServicesError>;
 }
 
-pub struct TcnMatcherImpl {}
-impl TcnMatcher for TcnMatcherImpl {
-  fn match_reports(&self, tcns: Vec<ObservedTcn>, reports: Vec<SignedReport>) -> Result<Vec<MatchedReport>, ServicesError> {
-    Self::match_reports_with(tcns, reports)
-  }
-}
- 
 #[derive(Debug, Clone)]
 pub struct MatchedReport{ report: SignedReport, contact_time: UnixTime }
-
-// TODO remove duplicate matcher functions from lib.rs
-impl TcnMatcherImpl {
-
-  // TODO use TCN repo's match_btreeset test code? Compare performance.
-  pub fn match_reports_with(tcns: Vec<ObservedTcn>, reports: Vec<SignedReport>) -> Result<Vec<MatchedReport>, ServicesError> {
-    let mut out: Vec<MatchedReport> = Vec::new();
-
-    let observed_tcns_map: HashMap<[u8; 16], ObservedTcn> = tcns.into_iter().map(|e|
-      (e.tcn.0, e)
-    ).collect();
-
-    for report in reports {
-      // TODO no unwrap
-      let rep = report.clone().verify().unwrap();
-      for tcn in rep.temporary_contact_numbers() {
-        if let Some(entry) = observed_tcns_map.get(&tcn.0) {
-          out.push(MatchedReport { report, contact_time: entry.time.clone() });
-          break;
-        }
-      }
-    }
-
-    Ok(out)
-  }
-}
-
-
-pub struct TcnMatcherOptimized {}
-
-impl TcnMatcher for TcnMatcherOptimized {
-  fn match_reports(&self, tcns: Vec<ObservedTcn>, reports: Vec<SignedReport>) -> Result<Vec<MatchedReport>, ServicesError> {
-    Self::match_reports_with(tcns, reports)
-  }
-}
- 
-impl TcnMatcherOptimized {
-
-  pub fn match_reports_with(tcns: Vec<ObservedTcn>, reports: Vec<SignedReport>) -> Result<Vec<MatchedReport>, ServicesError> {
-    let observed_tcns_map: HashMap<[u8; 16], ObservedTcn> = tcns.into_iter().map(|e|
-      (e.tcn.0, e)
-    ).collect();
-
-    let observed_tcns_map = Arc::new(observed_tcns_map);
-
-    let threads: Vec<JoinHandle<Option<MatchedReport>>> = reports.into_iter().map(|report| {
-      let observed_tcns_map = observed_tcns_map.clone();
-      thread::spawn(move || Self::match_report_with(&observed_tcns_map, report)
-    )}).collect();
-
-    let res: Vec<MatchedReport> = threads
-      .into_iter()
-      .map(|t| t.join().unwrap())
-      .filter_map(|option| option) // drop None (reports that didn't match)
-      .collect();
-
-    Ok(res)
-  }
-
-  pub fn match_report_with(observed_tcns_map: &HashMap<[u8; 16], ObservedTcn>, report: SignedReport) -> Option<MatchedReport> {
-    let mut out: Option<MatchedReport> = None;
-
-    let report_str = base64::encode(signed_report_to_bytes(report.clone()));
-
-    // TODO no unwrap
-    let rep = report.clone().verify().unwrap();
-    for tcn in rep.temporary_contact_numbers().take(10) {
-      if let Some(entry) = observed_tcns_map.get(&tcn.0) {
-        out = Some(MatchedReport { report, contact_time: entry.time.clone() });
-        break;
-      }
-    }
-    out
-  }
-}
-
-fn signed_report_to_bytes(signed_report: SignedReport) -> Vec<u8> {
-  let mut buf = Vec::new();
-  signed_report.write(Cursor::new(&mut buf)).expect("Couldn't write signed report bytes");
-  buf
-}
-
-
 
 pub struct TcnMatcherRayon {}
 
@@ -150,6 +60,53 @@ impl TcnMatcherRayon {
     out
   }
 }
+
+pub struct TcnMatcherStdThreadSpawn {}
+
+impl TcnMatcher for TcnMatcherStdThreadSpawn {
+  fn match_reports(&self, tcns: Vec<ObservedTcn>, reports: Vec<SignedReport>) -> Result<Vec<MatchedReport>, ServicesError> {
+    Self::match_reports_with(tcns, reports)
+  }
+}
+ 
+impl TcnMatcherStdThreadSpawn {
+
+  pub fn match_reports_with(tcns: Vec<ObservedTcn>, reports: Vec<SignedReport>) -> Result<Vec<MatchedReport>, ServicesError> {
+    let observed_tcns_map: HashMap<[u8; 16], ObservedTcn> = tcns.into_iter().map(|e|
+      (e.tcn.0, e)
+    ).collect();
+
+    let observed_tcns_map = Arc::new(observed_tcns_map);
+
+    let threads: Vec<JoinHandle<Option<MatchedReport>>> = reports.into_iter().map(|report| {
+      let observed_tcns_map = observed_tcns_map.clone();
+      thread::spawn(move || Self::match_report_with(&observed_tcns_map, report)
+    )}).collect();
+
+    let res: Vec<MatchedReport> = threads
+      .into_iter()
+      .map(|t| t.join().unwrap())
+      .filter_map(|option| option) // drop None (reports that didn't match)
+      .collect();
+
+    Ok(res)
+  }
+
+  pub fn match_report_with(observed_tcns_map: &HashMap<[u8; 16], ObservedTcn>, report: SignedReport) -> Option<MatchedReport> {
+    let mut out: Option<MatchedReport> = None;
+
+    // TODO no unwrap
+    let rep = report.clone().verify().unwrap();
+    for tcn in rep.temporary_contact_numbers().take(10) {
+      if let Some(entry) = observed_tcns_map.get(&tcn.0) {
+        out = Some(MatchedReport { report, contact_time: entry.time.clone() });
+        break;
+      }
+    }
+    out
+  }
+}
+
 
 #[derive(Debug, Eq, PartialEq, Clone)]
 pub struct ObservedTcn { tcn: TemporaryContactNumber, time: UnixTime }
@@ -465,7 +422,7 @@ struct SignedReportsChunk {
 mod tests {
   use super::*;
   use tcn::{MemoType, ReportAuthorizationKey};
-  use crate::reporting::memo::MemoMapperImpl;
+  use crate::reporting::{symptom_inputs::UserInput, memo::MemoMapperImpl, public_report::{CoughSeverity, FeverSeverity}};
   use std::io::Cursor;
 
   #[test]
@@ -481,19 +438,6 @@ mod tests {
     assert_eq!(observed_tc_from_bytes, observed_tcn);
   }
 
-  fn create_test_report() -> SignedReport {
-    let memo_mapper = MemoMapperImpl {};
-    let public_report = PublicReport {
-      earliest_symptom_time: UserInput::Some(UnixTime { value: 1589209754 }),
-      fever_severity: FeverSeverity::Serious,
-      breathlessness: true,
-      cough_severity: CoughSeverity::Existing
-    };
-    let rak = ReportAuthorizationKey::new(rand::thread_rng());
-    let memo_data = memo_mapper.to_memo(public_report, UnixTime { value: 1589209754 });
-    rak.create_report(MemoType::CoEpiV1, memo_data.bytes, 1, 10000).unwrap()
-  }
-
   #[test]
   fn matching_benchmark() {
 
@@ -502,11 +446,10 @@ mod tests {
     let verification_contact_time =  UnixTime { value: 1590528300 };
     let verification_report = SignedReport::with_str(verification_report_str).unwrap();
 
-    let mut reports: Vec<SignedReport> = vec![0; 10000].into_iter().map(|_| create_test_report()).collect();
+    let mut reports: Vec<SignedReport> = vec![0; 20].into_iter().map(|_| create_test_report()).collect();
     reports.push(verification_report);
 
-    // let matcher = TcnMatcherImpl {}; // 20 -> 6s, 200 -> 64s, 1000 -> 317s, 10000 ->
-    // let matcher = TcnMatcherOptimized {}; // 20 -> 1s, 200 -> 16s, 1000 -> 84s, 10000 ->
+    // let matcher = TcnMatcherStdThreadSpawn {}; // 20 -> 1s, 200 -> 16s, 1000 -> 84s, 10000 ->
     let matcher = TcnMatcherRayon {}; // 20 -> 1s, 200 -> 7s, 1000 -> 87s, 10000 -> 927s
 
     let tcns = vec![
@@ -529,5 +472,24 @@ mod tests {
     let matched_report_str = base64::encode(signed_report_to_bytes(matches[0].report.clone()));
     assert_eq!(matched_report_str, verification_report_str);
     assert_eq!(matches[0].contact_time, verification_contact_time);
+  }
+
+  fn create_test_report() -> SignedReport {
+    let memo_mapper = MemoMapperImpl {};
+    let public_report = PublicReport {
+      earliest_symptom_time: UserInput::Some(UnixTime { value: 1589209754 }),
+      fever_severity: FeverSeverity::Serious,
+      breathlessness: true,
+      cough_severity: CoughSeverity::Existing
+    };
+    let rak = ReportAuthorizationKey::new(rand::thread_rng());
+    let memo_data = memo_mapper.to_memo(public_report, UnixTime { value: 1589209754 });
+    rak.create_report(MemoType::CoEpiV1, memo_data.bytes, 1, 10000).unwrap()
+  }
+
+  fn signed_report_to_bytes(signed_report: SignedReport) -> Vec<u8> {
+    let mut buf = Vec::new();
+    signed_report.write(Cursor::new(&mut buf)).expect("Couldn't write signed report bytes");
+    buf
   }
 }