Added 2 approaches:
- `std::thread::spawn` (no thread pool)
- Using [rayon](https://github.com/rayon-rs/rayon) crate

Benchmarks so far (tested on my computer):
3 stored TCNs
all reports have j1=1, j2=10000
report count -> time
- Default implementation: 
20 -> 6s, 200 -> 64s, 1000 -> 317s, 10000 -> ?
- `std::thread::spawn`: 
20 -> 1s, 200 -> 16s, 1000 -> 84s, 10000 -> ?
- rayon: 
20 -> 1s, 200 -> 7s, 1000 -> 87s, 10000 -> 927s